optimize: drop field_forward pass; extract alias helpers to optimize/alias.rs
Step 4 of integrating field_forward into const_fold (issue #1009).

* Remove the `tir/field_forward` step from the optimization loop —
  const_fold now subsumes its rewrite responsibilities through
  `Interpreter::field_env` and the visitor's branch-aware
  fork/restore.
* Move the alias / value-copy-helper analyses
  (`build_alias_info`, `build_value_copy_helpers`,
  `recognize_value_copy`, plus the supporting walks
  `collect_aliased_in_*`, `collect_alias_groups`,
  `expr_for_each_child`, `stmt_for_each_child`,
  `type_creates_alias`) into `optimize/alias.rs`.
* Delete the now-empty `optimize/field_forward.rs`.

`sqlite_parse` at `-O3 --optimize-inline-threshold 40` (issue
#1009's reproducer):
- before: non-convergent, ~80 s at the original 100-iter cap and
  ~28 s at PR #1005's 30-iter cap.
- after: converges in 5 iterations, 13.9 s with the default cap and
  13.8 s with `--optimize-iterations 100` (the cap is irrelevant
  once the loop reaches a true fixed point).

5765 e2e fixtures + 448 unit tests pass.
